+++ CHANGELOG.md
@@ -18,6 +18,11 @@
 - Rename `Context::add_log` to `Context::add_attribute`.
 - Add `Api::debug` for emitting debug messages during development.
 
+**cosmwasm-vm**
+
+- `CosmCache::new`, `Instance::from_code` and `Instance::from_module` now take
+  an additional argument to enable/disable printing debug logs from contracts.
+
 ## 0.10.1 (2020-08-25)
 
 **cosmwasm-std**